ACPI: generic initramfs table override support
This patch allows the system administrator to override ACPI tables with
versions provided in an initramfs.

This works by moving the initialization of populate_rootfs earlier
in the initialization so that we can use the VFS file system routines.
The system is initialized enough to support this by acpi_early_init().
My understanding is that an early version of original patch posted at
http://gaugusch.at/kernel.shtml may have done something similar.

This version provides the infrastructure to override any ACPI table, but
only provides support for overriding DSDT. If other tables are desired,
extending the support is trivial.

During early ACPI initialization, when the initramfs is still loaded,
we go through a table of override entries which specify the name of the
table to override, the file name that contains it, and a pointer to the
data loaded from the file. The override tables and headers are kept in
memory so that they available to the ACPI subsystem after the __init
sections and the initramfs have been jettisoned.

This patch is derived from the work by Éric Piel <eric.piel@tremplin-utc.net>.

13 Jan 2010 jeffm: Uses initramfs_{read,write} now to avoid
                   "scheduling while atomic" warnings.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>