sched-bfs: worker fix
Seems like some worker function in the wifi code in 2.6.36 hits this bug_on,
which it shouldn't. It's hard to know if this is caused by BFS per se, or is
in iwlagn itself. It is, however, harmless to wake up a task on another RQ
with BFS, so just disable this BUG_ON for now.

-ck